+++ browser/brave_wallet/BUILD.gn
@@ -84,7 +84,7 @@ source_set("tab_helper") {
 
   deps = [
     "//brave/common",
-    "//brave/components/brave_wallet/browser:browser",
+    "//brave/components/brave_wallet/browser",
     "//brave/components/brave_wallet/browser:ethereum_permission_utils",
     "//brave/components/resources:strings_grit",
     "//components/permissions",
