+++ components/brave_wallet/browser/brave_wallet_utils.cc
@@ -486,90 +486,6 @@ void UpdateLastUnlockPref(PrefService* prefs) {
   prefs->SetTime(kBraveWalletLastUnlockTime, base::Time::Now());
 }
 
-absl::optional<mojom::EthereumChain> ValueToEthereumChain(
-    const base::Value& value) {
-  mojom::EthereumChain chain;
-  const base::DictionaryValue* params_dict = nullptr;
-  if (!value.GetAsDictionary(&params_dict) || !params_dict)
-    return absl::nullopt;
-
-  const std::string* chain_id = params_dict->FindStringKey("chainId");
-  if (!chain_id) {
-    return absl::nullopt;
-  }
-  chain.chain_id = *chain_id;
-
-  const std::string* chain_name = params_dict->FindStringKey("chainName");
-  if (chain_name) {
-    chain.chain_name = *chain_name;
-  }
-
-  const base::Value* explorerUrlsListValue =
-      params_dict->FindListKey("blockExplorerUrls");
-  if (explorerUrlsListValue) {
-    for (const auto& entry : explorerUrlsListValue->GetList())
-      chain.block_explorer_urls.push_back(entry.GetString());
-  }
-
-  const base::Value* iconUrlsValue = params_dict->FindListKey("iconUrls");
-  if (iconUrlsValue) {
-    for (const auto& entry : iconUrlsValue->GetList())
-      chain.icon_urls.push_back(entry.GetString());
-  }
-
-  const base::Value* rpcUrlsValue = params_dict->FindListKey("rpcUrls");
-  if (rpcUrlsValue) {
-    for (const auto& entry : rpcUrlsValue->GetList())
-      chain.rpc_urls.push_back(entry.GetString());
-  }
-  const std::string* symbol_name = params_dict->FindStringKey("name");
-  if (symbol_name) {
-    chain.name = *symbol_name;
-  }
-  const std::string* symbol = params_dict->FindStringKey("symbol");
-  if (symbol) {
-    chain.symbol = *symbol;
-  }
-  absl::optional<int> decimals = params_dict->FindIntKey("decimals");
-  if (decimals) {
-    chain.decimals = decimals.value();
-  }
-  return chain;
-}
-
-base::Value EthereumChainToValue(const mojom::EthereumChainPtr& chain) {
-  base::Value dict(base::Value::Type::DICTIONARY);
-  dict.SetStringKey("chainId", chain->chain_id);
-  dict.SetStringKey("chainName", chain->chain_name);
-
-  base::ListValue blockExplorerUrlsValue;
-  if (!chain->block_explorer_urls.empty()) {
-    for (const auto& url : chain->block_explorer_urls) {
-      blockExplorerUrlsValue.AppendString(url);
-    }
-  }
-  dict.SetKey("blockExplorerUrls", std::move(blockExplorerUrlsValue));
-
-  base::ListValue iconUrlsValue;
-  if (!chain->icon_urls.empty()) {
-    for (const auto& url : chain->icon_urls) {
-      iconUrlsValue.AppendString(url);
-    }
-  }
-  dict.SetKey("iconUrls", std::move(iconUrlsValue));
-
-  base::ListValue rpcUrlsValue;
-  for (const auto& url : chain->rpc_urls) {
-    rpcUrlsValue.AppendString(url);
-  }
-  dict.SetKey("rpcUrls", std::move(rpcUrlsValue));
-
-  dict.SetStringKey("name", chain->name);
-  dict.SetStringKey("symbol", chain->symbol);
-  dict.SetIntKey("decimals", chain->decimals);
-  return dict;
-}
-
 base::Value TransactionReceiptToValue(const TransactionReceipt& tx_receipt) {
   base::Value dict(base::Value::Type::DICTIONARY);
   dict.SetStringKey("transaction_hash", tx_receipt.transaction_hash);
