+++ components/brave_wallet/browser/brave_wallet_utils.h
@@ -95,9 +95,6 @@ base::Value TransactionReceiptToValue(const TransactionReceipt& tx_receipt);
 absl::optional<TransactionReceipt> ValueToTransactionReceipt(
     const base::Value& value);
 
-base::Value EthereumChainToValue(const mojom::EthereumChainPtr& chain);
-absl::optional<mojom::EthereumChain> ValueToEthereumChain(
-    const base::Value& value);
 void GetAllKnownChains(std::vector<mojom::EthereumChainPtr>* chains);
 const std::vector<mojom::KnownNetwork> GetAllKnownNetworks();
 void GetAllCustomChains(PrefService* prefs,
